Merge commit '1693a9d39bf3728534555a5ca865bead831cbef0' into feature/image_metadata

Conflicts:
	engine/src/ibmp.cpp
	engine/src/igif.cpp
	engine/src/iimport.cpp
	engine/src/ijpg.cpp
	engine/src/image.h
	engine/src/image_rep.h
	engine/src/image_rep_densitymapped.cpp
	engine/src/image_rep_encoded.cpp
	engine/src/imageloader.cpp
	engine/src/imageloader.h
	engine/src/ipng.cpp

+++ engine/src/tilecachegl.cpp
@@ -16,7 +16,6 @@ along with LiveCode.  If not see <http://www.gnu.org/licenses/>.  */
 
 #include "prefix.h"
 
-#include "core.h"
 #include "globdefs.h"
 #include "filedefs.h"
 #include "objdefs.h"
@@ -136,7 +135,7 @@ static inline void glvColor4ub(MCTileCacheOpenGLVersion p_version, uint8_t r, ui
 
 ////////////////////////////////////////////////////////////////////////////////
 
-static inline void MCTileCacheOpenGLCompositorDecodeTile(MCTileCacheOpenGLCompositorContext *self, uint32_t p_tile, uint32_t& r_super_tile, uint32_t& r_sub_tile)
+static inline void MCTileCacheOpenGLCompositorDecodeTile(MCTileCacheOpenGLCompositorContext *self, uintptr_t p_tile, uint32_t& r_super_tile, uint32_t& r_sub_tile)
 {
 	r_super_tile = (p_tile & 0xffff) - 1;
 	r_sub_tile = (p_tile >> 16) - 1;
@@ -219,7 +218,7 @@ static bool MCTileCacheOpenGLCompositorCreateTile(MCTileCacheOpenGLCompositorCon
 	return true;
 }
 
-static void MCTileCacheOpenGLCompositorDestroyTile(MCTileCacheOpenGLCompositorContext *self, uint32_t p_tile)
+static void MCTileCacheOpenGLCompositorDestroyTile(MCTileCacheOpenGLCompositorContext *self, uintptr_t p_tile)
 {
 	// Fetch the super/sub indices of the tile.
 	uint32_t t_super_tile_index, t_sub_tile_index;
@@ -376,7 +375,7 @@ void MCTileCacheOpenGLCompositor_DeallocateTile(void *p_context, void *p_tile)
 	MCTileCacheOpenGLCompositorContext *self;
 	self = (MCTileCacheOpenGLCompositorContext *)p_context;
 	
-	MCTileCacheOpenGLCompositorDestroyTile(self, (uint32_t)p_tile);
+	MCTileCacheOpenGLCompositorDestroyTile(self, (uintptr_t)p_tile);
 }
 
 static void MCTileCacheOpenGLCompositor_PrepareFrame(MCTileCacheOpenGLCompositorContext *self)
@@ -541,12 +540,12 @@ bool MCTileCacheOpenGLCompositor_BeginLayer(void *p_context, const MCRectangle& 
 			glScissor(p_clip . x + self -> origin_x, p_clip . y + self -> origin_y, p_clip . width, p_clip . height);
 	}
 	
-	if (!self -> is_blending)
-	{
-		self -> is_blending = true;
-		glEnable(GL_BLEND);
-	}
-			
+    if (!self -> is_blending)
+    {
+        self -> is_blending = true;
+        glEnable(GL_BLEND);
+    }
+		
 	self -> current_opacity = p_opacity;
 	
 	return true;
@@ -570,7 +569,7 @@ bool MCTileCacheOpenGLCompositor_CompositeTile(void *p_context, int32_t p_x, int
 	self = (MCTileCacheOpenGLCompositorContext *)p_context;
 
 	uint32_t t_super_tile_index, t_sub_tile_index;
-	MCTileCacheOpenGLCompositorDecodeTile(self, (uint32_t)p_tile, t_super_tile_index, t_sub_tile_index);
+	MCTileCacheOpenGLCompositorDecodeTile(self, (uintptr_t)p_tile, t_super_tile_index, t_sub_tile_index);
 	
 	GLuint t_texture;
 	t_texture = self -> super_tiles[t_super_tile_index] -> texture;
